@@ -105,5 +25,110 @@
       "last 1 safari version"
     ]
   },
-  "proxy": "https://localhost:1234"
+  "keywords": [
+    "registrant",
+    "domain-management",
+    "react"
+  ],
+  "bugs": {
+    "url": "https://github.com/internetee/registrant_center/issues"
+  },
+  "homepage": "https://registrant.internet.ee",
+  "contributors": [
+    "Sergei Tsõganov <sergei.tsoganov@internet.ee> (https://github.com/maricavor)"
+  ],
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "dev:server": "cross-env NODE_ENV=development nodemon server/index.js --watch server",
+    "prod:server": "cross-env NODE_ENV=production nodemon server/index.js --watch server",
+    "start": "concurrently \"npm run dev:server\" \"npm run dev\" \"npm run lint:watch\" --kill-others --kill-others-on-fail",
+    "build": "node ./buildScript.js",
+    "serve": "cross-env NODE_ENV=production node server/index.js",
+    "lint": "eslint . --max-warnings=0",
+    "lint:fix": "eslint . --fix",
+    "lint:watch": "nodemon --watch ./src --watch ./server --ext js,jsx --exec \"npm run lint\"",
+    "serve:test": "cross-env NODE_ENV=test nodemon server/index.js",
+    "test": "npm run lint && vitest run",
+    "coverage": "vitest run --coverage",
+    "test:e2e": "concurrently \"npm run serve:test\" \"cypress run\" --kill-others --success first",
+    "cypress:open": "concurrently \"npm run serve:test\" \"cypress open\" --kill-others",
+    "cypress:run": "concurrently \"npm run serve:test\" \"cypress run\" --kill-others --success first",
+    "deploy": "npm run test && npm run build && npm run test:e2e && npm run serve",
+    "preview": "vite preview --port 3000",
+    "clean-logs": "rm -rf logs && mkdir logs"
+  },
+  "dependencies": {
+    "@reduxjs/toolkit": "^2.5.0",
+    "@vitejs/plugin-react": "^4.3.4",
+    "axios": "^1.7.9",
+    "body-parser": "^1.20.3",
+    "capitalize": "^2.0.4",
+    "classnames": "^2.5.1",
+    "compression": "^1.7.5",
+    "cookie-parser": "^1.4.7",
+    "cookie-session": "^2.1.0",
+    "core-js": "^3.39.0",
+    "cross-env": "^7.0.3",
+    "dompurify": "^3.2.3",
+    "dotenv": "^16.4.7",
+    "eis-registrant-theme": "git+https://github.com/internetee/registrant_center_theme.git",
+    "express": "^4.21.2",
+    "express-winston": "^4.2.0",
+    "fs-extra": "^11.2.0",
+    "grant-express": "^5.2.0",
+    "helmet": "^8.0.0",
+    "jsonwebtoken": "^9.0.2",
+    "jwk-to-pem": "^2.0.7",
+    "pdfmake": "^0.2.15",
+    "prop-types": "^15.8.1",
+    "react": "18.2.0",
+    "react-cookie": "^7.2.2",
+    "react-csv": "^2.2.2",
+    "react-dom": "18.2.0",
+    "react-flatpickr": "^3.10.13",
+    "react-ga": "^3.3.1",
+    "react-helmet-async": "^2.0.5",
+    "react-intl": "^7.0.4",
+    "react-layout-masonry": "^1.2.0",
+    "react-redux": "^9.2.0",
+    "react-responsive": "^10.0.0",
+    "react-router-dom": "^7.1.1",
+    "redux": "^5.0.1",
+    "redux-state-sync": "^3.1.4",
+    "regenerator-runtime": "^0.14.1",
+    "sass": "^1.82.0",
+    "semantic-ui-react": "^3.0.0-beta.2",
+    "serve-favicon": "^2.5.0",
+    "universal-cookie": "^7.2.2",
+    "vanilla-cookieconsent": "^3.0.1",
+    "vite": "^6.0.1",
+    "winston": "^3.17.0",
+    "winston-daily-rotate-file": "^5.0.0"
+  },
+  "devDependencies": {
+    "@eslint/js": "^9.17.0",
+    "@testing-library/dom": "^10.4.0",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^16.1.0",
+    "@testing-library/user-event": "^14.5.2",
+    "@types/react": "^19.0.1",
+    "@types/react-dom": "^19.0.1",
+    "@vitest/coverage-v8": "^2.1.8",
+    "concurrently": "^9.1.0",
+    "cypress": "^13.17.0",
+    "eslint": "^8.57.1",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-import": "^2.31.0",
+    "eslint-plugin-jsx-a11y": "^6.10.2",
+    "eslint-plugin-prettier": "^5.2.1",
+    "eslint-plugin-react": "^7.37.2",
+    "eslint-plugin-react-hooks": "^4.6.2",
+    "eslint-plugin-react-refresh": "^0.4.16",
+    "globals": "^15.14.0",
+    "jsdom": "^25.0.1",
+    "nodemon": "^3.1.7",
+    "vitest": "^2.1.8"
+  }
 }
