@@ -1,27 +1,21 @@
-import React from 'react';
-import { Redirect, Route } from 'react-router-dom';
+import { Navigate } from 'react-router';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 
-const ProtectedRoute = ({ fetchUser, user, ...rest }) => {
-    const { data } = user;
-    const { name } = data;
-
-    if (!name) {
-        return <Redirect to="/login" />;
+const ProtectedRoute = ({ user, children }) => {
+    if (!user?.data?.name) {
+        return <Navigate replace to="/login" />;
     }
-    return <Route {...rest} />;
+    return children;
 };
 
-const mapStateToProps = ({ user }) => ({ user });
+const mapStateToProps = (state) => ({
+    user: state.user,
+});
 
 export default connect(mapStateToProps)(ProtectedRoute);
 
 ProtectedRoute.propTypes = {
-    fetchUser: PropTypes.string,
     user: PropTypes.object.isRequired,
-};
-
-ProtectedRoute.defaultProps = {
-    fetchUser: '',
+    children: PropTypes.node.isRequired,
 };
