@@ -1,19 +1,21 @@
 /* eslint-disable no-underscore-dangle */
 import dotenv from 'dotenv';
+dotenv.config();
+
 import axios from 'axios';
 import qs from 'qs';
 import jwt from 'jsonwebtoken';
 import capitalize from 'capitalize';
-
-dotenv.config();
+import { logDebug, logInfo, logError } from '../utils/logger.js';
 
 const {
     CLIENT_ID,
     CLIENT_SECRET,
     HOST,
+    PORT,
     ISSUER_URL,
     NODE_ENV,
-    REACT_APP_SERVER_PORT,
+    VITE_SERVER_PORT,
     REDIRECT_URL,
     TOKEN_PATH,
 } = process.env;
@@ -22,7 +24,7 @@ const B64_VALUE = Buffer.from(`${CLIENT_ID}:${CLIENT_SECRET}`).toString('base64'
 
 const redirect_uri =
     NODE_ENV === 'development'
-        ? `https://${HOST}:${REACT_APP_SERVER_PORT}${REDIRECT_URL}`
+        ? `https://${HOST}:${VITE_SERVER_PORT}${REDIRECT_URL}`
         : `https://${HOST}${REDIRECT_URL}`;
 
 export default async function callbackPageRoute(req, res, publicKey) {
@@ -31,12 +33,10 @@ export default async function callbackPageRoute(req, res, publicKey) {
             throw new Error(req.query.error);
         }
 
-        /* Võta päringu query-osast TARA poolt saadetud volituskood (authorization code) */
+        // Võta päringu query-osast EEID poolt saadetud volituskood (authorization code)
         const { code } = req.query;
 
-        /*
-     Turvaelemendi state kontroll
-    */
+        // Turvaelemendi state kontroll
         const returnedState = req.query.state;
         const sessionState = req.session.grant.state;
         if (returnedState !== sessionState) {
@@ -59,7 +59,7 @@ export default async function callbackPageRoute(req, res, publicKey) {
 
         const {
             data: { id_token },
-        } = await axios(options); // eslint-disable-line camelcase
+        } = await axios(options);
 
         /*
          Identsustõendi kontrollimine. Teegi jsonwebtoken
@@ -95,20 +95,24 @@ export default async function callbackPageRoute(req, res, publicKey) {
                     last_name: capitalize.words(verifiedJwt.profile_attributes.family_name),
                 };
 
-                console.log('Decrypted JWT from TARA:');
-                console.log(verifiedJwt);
+                logDebug('Decrypted JWT from TARA:', { verifiedJwt });
+                logInfo('User logged in successfully!');
+
                 req.session.user = userData;
                 if (NODE_ENV === 'development') {
-                    res.redirect(`https://${HOST}:3000`);
+                    res.redirect(`http://${HOST}:${PORT}`);
                 } else {
                     res.redirect('/');
                 }
             }
         );
     } catch (e) {
-        console.log(e); // eslint-disable-line no-console
+        logError('Authentication failed', e);
+        if (process.env.NODE_ENV !== 'test') {
+            console.warn(e);
+        }
         if (NODE_ENV === 'development') {
-            res.redirect(`https://${HOST}:3000/login`);
+            res.redirect(`http://${HOST}:${PORT}/login`);
         } else {
             res.redirect('/login');
         }
