@@ -1,7 +1,7 @@
-import React, { useEffect } from 'react';
+import { useEffect } from 'react';
 import { FormattedMessage } from 'react-intl';
 import { connect } from 'react-redux';
-import { Redirect, useParams } from 'react-router-dom';
+import { useNavigate, useParams } from 'react-router-dom';
 import { bindActionCreators } from 'redux';
 import PropTypes from 'prop-types';
 import { MainLayout, PageMessage } from '../../components';
@@ -11,16 +11,14 @@ const langs = ['et', 'en'];
 
 const ErrorPage = ({ setLang }) => {
     const { lang } = useParams();
+    const navigate = useNavigate();
 
     useEffect(() => {
         if (langs.includes(lang)) {
             setLang(lang);
+            navigate('/', { replace: true });
         }
-    }, [lang, setLang]);
-
-    if (langs.includes(lang)) {
-        return <Redirect to="/" />;
-    }
+    }, [lang, setLang, navigate]);
 
     return (
         <MainLayout hasBackButton titleKey="errorpage.title">
@@ -41,7 +39,6 @@ const mapDispatchToProps = (dispatch) =>
     );
 
 export default connect(null, mapDispatchToProps)(ErrorPage);
-
 ErrorPage.propTypes = {
     setLang: PropTypes.func.isRequired,
 };
