@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from 'react';
+import { useEffect, useState } from 'react';
 import { FormattedMessage, useIntl } from 'react-intl';
 import { Grid, Icon } from 'semantic-ui-react';
 import { Link } from 'react-router-dom';
@@ -28,7 +28,7 @@ const HomePage = ({
     const dispatch = useDispatch();
 
     const [techParams, setTechParams] = useState(isTech);
-    const [previousTechParams, setPreviousTechParams] = useState(isTech);
+    const [previousTechParams] = useState(isTech);
 
     const onSelectTech = (value) => {
         setIsLoading(true);
@@ -50,7 +50,7 @@ const HomePage = ({
         } else {
             setIsLoading(false);
         }
-    }, [fetchDomains, isTech]);
+    }, [fetchDomains, isTech, domains.length, previousTechParams]);
 
     if (isLoading) return <Loading />;
 
@@ -105,8 +105,8 @@ const HomePage = ({
                 </div>
                 <DomainList
                     domainCount={totalDomains}
-                    domains={domains}
                     domainTotal={absoluteCount}
+                    domains={domains}
                     isTech={isTech}
                     isUpdateContact={isUpdateContact}
                     lang={lang}
