@@ -1,16 +1,28 @@
-import React from 'react';
 import { FormattedMessage } from 'react-intl';
 import { connect } from 'react-redux';
 import { Button, Container, Icon, Form } from 'semantic-ui-react';
 import PropTypes from 'prop-types';
 import { MainLayout, MessageModule } from '../../components';
 
-const { REACT_APP_SCOPE, REACT_APP_SERVER_PORT, REACT_APP_URL } = process.env;
+const { VITE_SERVER_PORT, VITE_URL } = import.meta.env;
 
-const authPath =
-    process.env.NODE_ENV === 'development'
-        ? `${REACT_APP_URL}:${REACT_APP_SERVER_PORT}/connect/openid`
-        : `${REACT_APP_URL}/connect/openid`;
+const getAuthUrl = (lang) => {
+    if (!VITE_URL || (process.env.NODE_ENV === 'development' && !VITE_SERVER_PORT)) {
+        console.error('Missing required environment variables for auth URL');
+        return null;
+    }
+
+    const baseUrl =
+        process.env.NODE_ENV === 'development' ? `${VITE_URL}:${VITE_SERVER_PORT}` : VITE_URL;
+
+    try {
+        // Ensure URL is properly encoded
+        return new URL(`/connect/openid/${encodeURIComponent(lang)}`, baseUrl).toString();
+    } catch (error) {
+        console.error('Failed to construct auth URL:', error);
+        return null;
+    }
+};
 
 function LoginPage({ user, ui }) {
     const message = {
@@ -18,8 +30,21 @@ function LoginPage({ user, ui }) {
         type: 'logout',
     };
 
+    const handleLogin = (e) => {
+        e.preventDefault();
+        const authUrl = getAuthUrl(ui.lang);
+
+        if (!authUrl) {
+            // You might want to show an error message to the user here
+            console.error('Unable to proceed with login - invalid auth URL');
+            return;
+        }
+
+        window.location.href = authUrl;
+    };
+
     return (
-        <MainLayout heroKey="login.hero.text" titleKey="login.title">
+        <MainLayout titleKey="login.title">
             {user.isLoggedOut && message.code && <MessageModule message={message} />}
             <div className="page page--login">
                 <div className="u-container">
@@ -36,28 +61,18 @@ function LoginPage({ user, ui }) {
                                 <FormattedMessage id="login.options.mobileId" tagName="p" />
                             </div>
                             <div className="login-options--item">
-                                <Icon name="lock" size="big" />
-                                <FormattedMessage id="login.options.eidas" tagName="p" />
-                            </div>
-                            <div className="login-options--item">
                                 <Icon name="mobile alternate" size="big" />
                                 <FormattedMessage id="login.options.smartId" tagName="p" />
                             </div>
+                            <div className="login-options--item">
+                                <Icon name="lock" size="big" />
+                                <FormattedMessage id="login.options.eidas" tagName="p" />
+                            </div>
                         </div>
-                        <Form action={authPath}>
+                        <Form onSubmit={handleLogin}>
                             <Button primary size={ui.uiElemSize} type="submit">
                                 <FormattedMessage id="actions.login" tagName="span" />
                             </Button>
-                            {REACT_APP_SCOPE && REACT_APP_SCOPE.includes('webauthn') && (
-                                <Button
-                                    formAction={`${authPath}/webauthn`}
-                                    primary
-                                    size={ui.uiElemSize}
-                                    type="submit"
-                                >
-                                    <FormattedMessage id="actions.webauthn-login" tagName="span" />
-                                </Button>
-                            )}
                         </Form>
                     </Container>
                 </div>
