@@ -1,2 +1,2 @@
 import 'core-js/stable';
-import 'regenerator-runtime/runtime';
+import 'regenerator-runtime';
